chore: Update go-api-http-client and go-api-sdk-jamfpro dependencies
This commit updates the go-api-http-client dependency to version 0.1.41 and the go-api-sdk-jamfpro dependency to version 1.6.10. These updates ensure that the latest stable versions of the dependencies are being used, improving the overall stability and functionality of the code.

Note: The recent user commits and repository commits are not directly related to the code changes provided.