feat: Update Jamf Pro resource documentation
This commit updates the documentation for the Jamf Pro resources, specifically the `policy.md`, `macos_configuration_profile.md`, and `macos_configuration_profile_plist.md` files. It adds a new read-only property `name` to the schema, which represents the name of the Jamf Pro Site based on the utilized site ID. This change improves the clarity and completeness of the resource documentation.