feat: Add script to export Jamf Pro user account privileges
This commit adds a new Go script that exports user account privileges from Jamf Pro. The script uses the go-api-sdk-jamfpro package to interact with the Jamf Pro API. It fetches the Jamf Pro version, creates a directory with the version number, creates a new account with full access privileges, retrieves the full account details, exports the privileges to JSON files, and finally deletes the created account. This script will be used for maintaining and updating the API privileges data.